rename JobOutputDir to JobOutputProject
Changes to model.py renaming JobOutputDir to JobOutputProject.
Removes dir_name field since it is no longer needed.
This model stores the project that is shared with the user and contains the output files.
The user needs to create this object so they can pick which DukeDS server the project
will live on. This will only be needed for end users if there are more than one DukeDS instance.