Adds ShareGroup and DDSUser tables/endpoint
These tables specify the DukeDS users who need to review the results
of a job that has been run. Admin users are expected to setup these tables.
The ShareGroup will be tied to the JobQuestionnaire.
When the Job table is created is should copy the ShareGroup value.

ShareGroup has a ManyToMany relationship with DDSUsers.
JobQuestionnaire has a OneToOne relationship with ShareGroup.
Job has a OneToOne relationship with ShareGroup.

Adds share_group to `admin/jobs` endpoint serializer as an integer.
Adds `/admin/share-groups` endpoint for looking up the list of users.